user

+++ app/Http/Controllers/API/UserController.php
@@ -8,6 +8,8 @@ use Illuminate\Http\Request;
 use Illuminate\Http\UploadedFile;
 use Illuminate\Support\Facades\Storage;
 use File;
+use Illuminate\Validation\Rule;
+use Auth;
 
 class UserController extends BaseController
 {
@@ -22,18 +24,24 @@ class UserController extends BaseController
 
     public function update(Request $request)
     {
-        $user = $request->user();
+        // $user = $request->user();
+        // dd(Auth::id());
         $validator = \Validator::make($request->all(), [
             'first_name' => 'required',
             // 'phone' => 'required|unique:users,phone,'. $request->user()->id,
-            'email' => 'required|email|unique:users,email,'. $request->user()->id
+            // 'email' => 'required|email|unique:users,email,'. $request->user()->id
+            'email' => [
+                'required',
+                'email',
+                Rule::unique('users')->ignore(Auth::id(), 'id')
+            ]
         ]);
 
         if($validator->fails()) {
             return $this->sendError('Error ', $validator->errors());
         }
 
-        $user   = User::findOrFail($user->id);
+        $user               = User::findOrFail(Auth::id());
         $image_old          = $user->image;
         $image              = $request->image;
         
@@ -45,9 +53,11 @@ class UserController extends BaseController
                 if ($request->hasFile('image')) {
                     $image_path = "/images/user/{$image_old}";
                     if ($request->file('image')->isValid()) {
-                        if (file_exists(public_path().$image_path)) {
-                           unlink(public_path().$image_path);
-                        }
+                        if($image_old != null || $image_old != 'NULL'){
+                            if (file_exists(public_path().$image_path)) {
+                               unlink(public_path().$image_path);
+                            }
+                        }                        
                         $getimageName = time().'.'.$request->image->getClientOriginalExtension();
                         $image = $request->image->move(public_path('images/user'), $getimageName);
                     }
